Be more selective in HeadsUp notifications
This will limit HeadsUp notifications somewhat.  Much of the logic
was pre-determined by Google, but I had bypassed it as the 'Threshold'
priority doesn't appear to be fully implemented.  However, other logic
factors such as 'noisy' and isAllowed  should be followed.
At minumum, the 'isNoisy' flag will limit the HeadsUp notification to
those that make a sound or vibrate (ie, it will limit the ongoing
notifications).

Signed-off-by: Mike Wilson <mwilson@acesappraisal.com>